add validation

+++ internal/book-service/books.go
@@ -9,16 +9,28 @@ import (
 )
 
 func (s *Service) Books(ctx context.Context, f bm.BookFilter) ([]bm.Book, error) {
-	if f.OrderBy == "" {
+	switch f.OrderBy {
+	case "id", "title", "author", "genre", "published_date", "edition":
+	case "":
 		f.OrderBy = "id"
+	default:
+		return nil, bm.ValidationError("incorrect order_by")
+	}
+
+	if f.Page < 0 {
+		return nil, bm.ValidationError("incorrect page")
 	}
 
 	if f.Page == 0 {
 		f.Page = 1
 	}
 
-	if f.PageSize == 0 {
-		f.PageSize = 50
+	if f.PageSize < 0 {
+		return nil, bm.ValidationError("page_size is negative")
+	}
+
+	if f.PageSize == 0 || f.PageSize > maxPageSize {
+		f.PageSize = maxPageSize
 	}
 
 	books, err := s.db.Books(ctx, f)
@@ -29,25 +41,53 @@ func (s *Service) Books(ctx context.Context, f bm.BookFilter) ([]bm.Book, error)
 	return books, nil
 }
 
-func (s *Service) CreateBook(ctx context.Context, book bm.Book) (int64, error) {
-	if !book.PublishedDate.IsZero() {
-		book.PublishedDate = book.PublishedDate.Truncate(24 * time.Hour)
+func (s *Service) CreateBook(ctx context.Context, b bm.Book) (int64, error) {
+	if b.Title == "" {
+		return 0, bm.ValidationError("title is empty")
 	}
 
-	ids, err := s.db.CreateBook(ctx, book)
+	if b.Author == "" {
+		return 0, bm.ValidationError("author is empty")
+	}
+
+	if b.Genre == "" {
+		return 0, bm.ValidationError("genre is empty")
+	}
+
+	if !b.PublishedDate.IsZero() {
+		b.PublishedDate = b.PublishedDate.Truncate(24 * time.Hour)
+	}
+
+	id, err := s.db.CreateBook(ctx, b)
 	if err != nil {
 		return 0, fmt.Errorf("create book: %w", err)
 	}
 
-	return ids, nil
+	return id, nil
 }
 
-func (s *Service) UpdateBook(ctx context.Context, book bm.Book) error {
-	if !book.PublishedDate.IsZero() {
-		book.PublishedDate = book.PublishedDate.Truncate(24 * time.Hour)
+func (s *Service) UpdateBook(ctx context.Context, b bm.Book) error {
+	if b.ID <= 0 {
+		return bm.ValidationError("incorrect id")
 	}
 
-	if err := s.db.UpdateBook(ctx, book); err != nil {
+	if b.Title == "" {
+		return bm.ValidationError("title is empty")
+	}
+
+	if b.Author == "" {
+		return bm.ValidationError("author is empty")
+	}
+
+	if b.Genre == "" {
+		return bm.ValidationError("genre is empty")
+	}
+
+	if !b.PublishedDate.IsZero() {
+		b.PublishedDate = b.PublishedDate.Truncate(24 * time.Hour)
+	}
+
+	if err := s.db.UpdateBook(ctx, b); err != nil {
 		return fmt.Errorf("update book: %w", err)
 	}
 
@@ -55,6 +95,10 @@ func (s *Service) UpdateBook(ctx context.Context, book bm.Book) error {
 }
 
 func (s *Service) DeleteBooks(ctx context.Context, ids []int64) error {
+	if len(ids) == 0 {
+		return bm.ValidationError("ids list is empty")
+	}
+
 	if err := s.db.DeleteBooks(ctx, ids); err != nil {
 		return fmt.Errorf("delete books: %w", err)
 	}
